requirements.txt
opencv-python
openvino
openvino-dev
pytest
xdist
pyhamcrest